Change Freelex timeout from 60 sec to 20 sec.
Heroku enforces a 30 second timeout on our generating a response. We set a
timeout for Freelex that allows us time to process the response and return it
to our client within that constraint.

Before this change, timeout and connection delays were sometimes being
reported as different exceptions e.g. `ActionView::Template::Error because
the Heroku timeout would cause Rails to fail elsewhere in its processing.