Workaround bug where autocomplete search results have incorrect encoding
The response from the autocomplete service sets the HTTP header:

    Content-Type: text/html; charset=ISO-8859-1

but in reality the body string is UTF-8. This was causing autocomplete
searches which included macrons to fail.

For example, the results for search term "wha" include strings with
macrons e.g. "Whangārei". By inspection of the response body in a hex
editor the "lowercase a with macron" is 0xC481 which corresponds to the
UTF-8 encoding for that grapheme - see https://en.wikipedia.org/wiki/%C4%80

To work around this, we tell ruby to change the encoding tag on the body
String (#force_encoding only changes the tag, it does not attempt to
transcode the data).