Make production logs more manageable.
Before this change we were logging very verbosely in the `production`
env e.g. a single request to  `GET /numbers` generated 462 lines of
logging output. This made it very difficult to find anything in the
logs.

This change does two things:

1. Change production log level from `debug` to `info`. This does not do
   much on its own - Rails still logs very verbosely at the `info` level
   and a single `GET /numbers` generated 450 lines of log output.
2. Install and configure the lograge gem. Lograge
   (https://github.com/roidrage/lograge) allows us to generate one line
   of logging output per successful request (stacktraces are still show
   as they would at the current Rails log level).
3. Tell lograge to ignore requests to `SignImageController#show` because
   that endpoint is serving all image assets so gets hit many more times
   than any other endpoint (e.g. `GET /numbers` generates 90 subsequent
   requests to `SignImageController#show`

The end result is that we get one concise line per requests.