Install Rack::CanonicalHost middleware to redirect non-canonical hostnames
This application has multiple DNS names for historical reasons. For SEO and
effective CDN caching purposes we want to have just one canonical DNS name but
we also cannot break existing links which might use the other names.

Traditionally this problem would be solved via URL rewriting in
Apache/Nginx/Varnish etc. but that is not possible because we are hosted on
Heroku.

The `rack-canonical-host` middleware allows us to solve this problem in Rails.
It redirects requests for non-canonical domain names to the canonical domain
name.

`Rack::Sendfile` is the first middleware in the Rails middleware stack (run
`bundle exec rails middleware` for details) so we are inserting the
Rack::CanonicalHost middleware right at the start. This means we don't waste
time processing a request which will be redirected by Rack::CanonicalHost
anyway.

More details: https://github.com/tylerhunt/rack-canonical-host