Set Cache-Control header for assets
Before this change, this app serves assets without a `Cache-Control`
header. The absence of this header causes the CDN to re-request assets
very regularly. We can safely set the `max-age` to 1 year because our
CSS, JS, Image assets are served as fingerprinted filenames by the asset
pipeline i.e.  the filename will change if the file content changes and
this will automatically bust the cache.

The typical Rails production setup uses Apache or Nginx to serve file
assets, both of which will set appropriate cache headers. We are hosted
on Heroku so we serve assets from the Rails app and must set caching
headers ourselves.

This setting is already enabled in other environments. This change
adds it to the production environment.