Cache "sign of the day" for 24 hours
This change makes the app cache the "sign of the day" shown on the homepage
for 24 hours. This will greatly speed up the most popular page on the site.

The "Show me another sign" button is now implemented by a separate controller
endpoint (route is `/random_sign`) - this will allow it to be tracked in
Google Analytics as a page view.