Recover from unexpected cache errors
Rails is pretty good at dealing with situations where the cache goes away
unexpectedly but just to be safe we are rescuing errors from cache lookups
and recovering by fetching the data directly. We also log to Raygun so
that these cache errors will be noticed if they are an issue.