Cache VocabSheet#items to mitigate performance regression
When we stopped implementing VocabSheet#items with an ActiveRecord relation
we stopped getting the benefit of the various layers of caching that AR
does for us. This caused a performance regression because

1. VocabSheet#items is called multiple times per request if the user has
  created a vocab sheet.
2. Every call to VocabSheet#items was hitting Freelex once for each item
   (So N items in a vocab sheet would result in N request to Freelex)

This change mitigates that performance regression by caching the items for
use later in generating the same response. It makes use of the ActiveRecord
`*_changed?` methods introduced in Rails 5.