Allow caching of Signs used for Item building in Rails cache
This change stores the Sign objects used to build vocab sheet Item objects
in Rails cache.

This feature can be toggled on/off and configured at runtime via an
environment variable:

  FEATURE_CACHE_ITEMS_ENABLED=true # => enabled
  FEATURE_CACHE_ITEMS_NUM_HOURS=24 # default

The Sign objects used to build items are re-fetched on every request from
a user who has a vocab sheet. This set of sign objects is much more constrained
than the full set - the only Sign objects we will cache are those which
are part of a vocab sheet so I predict that we can cache Sign objects in
this case without overwhelming our small in-memory cache allocation.

I am using a feature flag because it is impossible to be certain about
this prediction until we try it in production (with production load).