Cache all Sign objects in `Rails.cache`
This change causes all Signs from Freelex to be cached in `Rails.cache`
as JSON.

The cache key is the query string sent to the Freelex because this
string is unique for each set of search results returned.

The signs are cached for a configurable number of hours (default is 24).
The cache expiry is configurable at runtime because we have a limited
amount of cache space available and expiry will have to be tuned.