Remove SignImagesController and ImageProcessor
This service used to be a resize and cache image proxy. The resizing
has been removed, so now this is either a no-op or a net negative for
performance depending on network and IO performance compared to accessing
the file directly.

Both Freelex and Signbank store files in S3, accessible via a public unsigned URL. This means that we can link to these assets directly without
needing to perform any transformation.