Always update the dictionary when a process starts in a deployed environment
In development and test, we restart processes all the time - fetching the dictionary
database is unnecessary overhead, especially when we can update it on-demand.

In staging and production, processes tend to start and then stay running. For this reason,
it's best that we update the dictionary every time these processes start, rather than
waiting for other rules to be met.

We tried to handle this assuming the dictionary wouldn't exist, but it ends up baked into the Heroku
slug, since asset precompilation starts the app. This stops the update from running until it's
older than 4 weeks, which isn't ideal if we're trying to get a database fix in.